backend/requirements.txt.txt
Django==4.2
djangorestframework
django-cors-headers
Pillow
python-dotenv
